Seperating labels from generate
